wOrking on questions and activitues

+++ scripts/libs.sh
@@ -4,8 +4,12 @@ set -e
 
 mkdir -pv lib
 
-wget -O lib/hamcrest-core-1.3.jar \
-    http://central.maven.org/maven2/org/hamcrest/hamcrest-core/1.3/hamcrest-core-1.3.jar
+if [ ! -f lib/hamcrest-core-1.3.jar ]; then
+    wget -O lib/hamcrest-core-1.3.jar \
+        http://central.maven.org/maven2/org/hamcrest/hamcrest-core/1.3/hamcrest-core-1.3.jar
+fi
 
-wget -O lib/junit-4.12.jar \
-    http://central.maven.org/maven2/junit/junit/4.12/junit-4.12.jar
+if [ ! -f lib/junit-4.12.jar ]; then
+    wget -O lib/junit-4.12.jar \
+        http://central.maven.org/maven2/junit/junit/4.12/junit-4.12.jar
+fi
